Ezra fight (#14)
* update

added fireball capability

* comments

comments

* fireball shoots now

* fireball works

shouldnt line 25 work? Are the characters not kinimaticBodies?

* update

added fireball capability

* comments

comments

* fireball shoots now

* fireball works

shouldnt line 25 work? Are the characters not kinimaticBodies?

* Remove extra pig

* it works

* Update Fireball.gd to use position_hit(power, pos)

It'll probably be easiest to use position_hit(int power, Vector2 position). This calculates the hit trajectory based on the fireball's relative position. For the power value, I'd recommend a range between 100-500. Once we can get this tested while playing, we can normalize that into a more reasonable scale.

* Update BasicPlayer.gd (#8)

Updated the player script to merge the fireball and the knock-back elements. Should be good to go.

* update